requirements.txt
args==0.1.0
astroid==2.2.5
clint==0.5.1
coverage==4.5.3
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.5.2
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
isort==4.3.15
lazy-object-proxy==1.3.1
mamba==0.10
mccabe==0.6.1
psycopg2==2.7.6.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
whitenoise==4.1.2
wrapt==1.11.1

